Restructured NavigationBar and broke it into smaller components.
